requirements.txt
bcrypt==4.3.0
blinker==1.9.0
capstone==6.0.0a5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colored-traceback==0.4.2
cryptography==45.0.7
Flask==3.1.2
idna==3.10
intervaltree==3.1.0
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
paramiko==4.0.0
plumbum==1.9.0
psutil==7.0.0
pwntools==4.14.1
pycparser==2.23
pyelftools==0.32
Pygments==2.19.2
PyNaCl==1.6.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.5
ROPGadget==7.6
rpyc==6.0.2
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
unicorn==2.1.4
unix-ar==0.2.1
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
zstandard==0.24.0
